requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file requirements.txt
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   groq
    #   httpx
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==5.5.2
    # via
    #   google-auth
    #   streamlit
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via streamlit
colorama==0.4.6
    # via click
distro==1.9.0
    # via groq
filetype==1.2.0
    # via langchain-google-genai
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
google-ai-generativelanguage==0.6.18
    # via
    #   sds-rag-tasks (pyproject.toml)
    #   langchain-google-genai
google-api-core==2.25.1
    # via google-ai-generativelanguage
google-auth==2.40.3
    # via
    #   google-ai-generativelanguage
    #   google-api-core
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
greenlet==3.2.4
    # via sqlalchemy
groq==0.31.0
    # via langchain-groq
grpcio==1.74.0
    # via
    #   google-api-core
    #   grpcio-status
    #   qdrant-client
grpcio-status==1.74.0
    # via google-api-core
h11==0.16.0
    # via httpcore
h2==4.2.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   groq
    #   langsmith
    #   qdrant-client
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.25.0
    # via altair
jsonschema-specifications==2025.4.1
    # via jsonschema
langchain==0.3.27
    # via sds-rag-tasks (pyproject.toml)
langchain-core==0.3.74
    # via
    #   sds-rag-tasks (pyproject.toml)
    #   langchain
    #   langchain-google-genai
    #   langchain-groq
    #   langchain-text-splitters
langchain-google-genai==2.1.9
    # via sds-rag-tasks (pyproject.toml)
langchain-groq==0.3.7
    # via sds-rag-tasks (pyproject.toml)
langchain-text-splitters==0.3.9
    # via langchain
langsmith==0.4.13
    # via
    #   langchain
    #   langchain-core
markupsafe==3.0.2
    # via jinja2
narwhals==2.0.1
    # via altair
numpy==2.3.2
    # via
    #   pandas
    #   pydeck
    #   qdrant-client
    #   streamlit
orjson==3.11.1
    # via langsmith
packaging==25.0
    # via
    #   altair
    #   langchain-core
    #   langsmith
    #   streamlit
pandas==2.3.1
    # via streamlit
pillow==11.3.0
    # via streamlit
portalocker==3.2.0
    # via qdrant-client
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==6.31.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
    #   qdrant-client
    #   streamlit
pyarrow==21.0.0
    # via streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via
    #   groq
    #   langchain
    #   langchain-core
    #   langchain-google-genai
    #   langsmith
    #   qdrant-client
pydantic-core==2.33.2
    # via pydantic
pydeck==0.9.1
    # via streamlit
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via sds-rag-tasks (pyproject.toml)
pytz==2025.2
    # via pandas
pywin32==311
    # via portalocker
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-core
qdrant-client==1.15.1
    # via sds-rag-tasks (pyproject.toml)
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.4
    # via
    #   google-api-core
    #   langchain
    #   langsmith
    #   requests-toolbelt
    #   streamlit
requests-toolbelt==1.0.0
    # via langsmith
rpds-py==0.27.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   groq
sqlalchemy==2.0.42
    # via langchain
streamlit==1.47.1
    # via sds-rag-tasks (pyproject.toml)
tenacity==9.1.2
    # via
    #   langchain-core
    #   streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.2
    # via streamlit
typing-extensions==4.14.1
    # via
    #   altair
    #   groq
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   qdrant-client
    #   requests
watchdog==6.0.0
    # via streamlit
zstandard==0.23.0
    # via langsmith
